Updated failsaf-plugin configuration to process **/*Test mask. Code clenup.
